Add files via upload
This Data Analysis includes the best time, day, and month to fly with minimum delays, the correlation between the older planes and delays, and the evolve in numbers of flights over time. As well as the construction of the model for the prediction of delays.
Data’s set used consists of flight arrival and departure details, from years 1998 to 1999 (HARVARD DATAVERSE, 2008). All Data are stored as data frame using SQLite and all analysis are done with RStudio.